#871 fix coverage and remove redundant test
